+++ Changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+## [v1.1.2]
 ### Added
 - pick outputs by serial number, name or make/brand.
 
@@ -62,7 +63,8 @@
 - position two displays relative to each other either left or right.
 - while positioning the displays, align them to bottom or top.
 
-[unreleased]: https://gitlab.com/eldoheiri/wltile/-/compare/v1.1.1...main
+[unreleased]: https://gitlab.com/eldoheiri/wltile/-/compare/v1.1.2...main
+[v1.1.2]: https://gitlab.com/eldoheiri/wltile/-/compare/v1.1.1...main
 [v1.1.1]: https://gitlab.com/eldoheiri/wltile/-/compare/v1.1.0...v1.1.1
 [v1.1.0]: https://gitlab.com/eldoheiri/wltile/-/compare/v1.0.0...v1.1.0
 [v1.0.0]: https://gitlab.com/eldoheiri/wltile/-/compare/v0.3.4...v1.0.0